# Replace `DigestDelegate` with inline functions