--- a/library/digest/api/digest.api
+++ b/library/digest/api/digest.api
@@ -31,6 +31,6 @@ public abstract class org/kotlincrypto/core/digest/Digest : java/security/Messag
 }
 
 public abstract class org/kotlincrypto/core/digest/internal/DigestState {
-	public synthetic fun <init> (Ljava/lang/String;IILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public synthetic fun <init> (Ljava/lang/String;IIJLkotlin/jvm/internal/DefaultConstructorMarker;)V
 }