… library/digest/src/commonMain/kotlin/org/kotlincrypto/core/digest/internal/DigestDelegate.kt