Closes #65