@05nelsonm 2024-12-19
Benchmark results:

<details>
    <summary>macOS</summary>

Before
```
jvm summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  33.326 ± 3.717  ns/op
DigestBenchmark.update  avgt    5  30.059 ± 1.304  ns/op

nativeHost summary:
Benchmark               Mode  Cnt   Score    Error  Units
DigestBenchmark.copy    avgt    5  82.635 ±  0.157  ns/op
DigestBenchmark.update  avgt    5  94.151 ± 11.237  ns/op

js summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5   96.879 ± 3.083  ns/op
DigestBenchmark.update  avgt    5  238.214 ± 2.161  ns/op

wasmJs summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  95.971 ± 0.270  ns/op
DigestBenchmark.update  avgt    5  70.961 ± 3.161  ns/op
```

After
```
jvm summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  13.947 ± 0.449  ns/op
DigestBenchmark.update  avgt    5  23.803 ± 1.937  ns/op

nativeHost summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  53.958 ± 1.742  ns/op
DigestBenchmark.update  avgt    5  54.831 ± 2.544  ns/op

Benchmark               Mode  Cnt    Score    Error  Units
DigestBenchmark.copy    avgt    5  103.703 ± 11.876  ns/op
DigestBenchmark.update  avgt    5  101.945 ±  0.864  ns/op

wasmJs summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  92.182 ± 3.005  ns/op
DigestBenchmark.update  avgt    5  59.581 ± 1.707  ns/op
```

</details>

<details>
    <summary>Linux</summary>

Before:
```
jvm summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  40.152 ± 0.782  ns/op
DigestBenchmark.update  avgt    5  46.597 ± 0.383  ns/op

nativeHost summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  129.570 ± 0.869  ns/op
DigestBenchmark.update  avgt    5   89.657 ± 0.313  ns/op

js summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  114.049 ± 2.524  ns/op
DigestBenchmark.update  avgt    5  260.435 ± 2.626  ns/op

wasmJs summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  125.157 ± 0.416  ns/op
DigestBenchmark.update  avgt    5  126.401 ± 0.402  ns/op
```

After:
```
jvm summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  17.413 ± 0.165  ns/op
DigestBenchmark.update  avgt    5  55.637 ± 0.418  ns/op

nativeHost summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  71.091 ± 0.269  ns/op
DigestBenchmark.update  avgt    5  76.284 ± 0.146  ns/op

js summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5   87.753 ± 1.120  ns/op
DigestBenchmark.update  avgt    5  149.459 ± 1.680  ns/op

wasmJs summary:
Benchmark               Mode  Cnt   Score   Error  Units
DigestBenchmark.copy    avgt    5  92.792 ± 0.365  ns/op
DigestBenchmark.update  avgt    5  79.619 ± 0.368  ns/op
```

</details>

<details>
    <summary>Windows</summary>

Before:
```
jvm summary:
Benchmark               Mode  Cnt   Score    Error  Units
DigestBenchmark.copy    avgt    5  55.117 � 35.923  ns/op
DigestBenchmark.update  avgt    5  48.785 �  0.442  ns/op

nativeHost summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  194.471 ± 1.146  ns/op
DigestBenchmark.update  avgt    5   78.197 ± 0.334  ns/op

js summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  116.433 ± 2.097  ns/op
DigestBenchmark.update  avgt    5  259.408 ± 2.821  ns/op

wasmJs summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  132.846 ± 0.127  ns/op
DigestBenchmark.update  avgt    5  133.631 ± 0.621  ns/op
```

After
```
jvm summary:
Benchmark               Mode  Cnt   Score    Error  Units
DigestBenchmark.copy    avgt    5  29.301 � 19.249  ns/op
DigestBenchmark.update  avgt    5  57.007 �  1.136  ns/op

nativeHost summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  139.888 ± 2.301  ns/op
DigestBenchmark.update  avgt    5   86.055 ± 0.165  ns/op

js summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5   99.265 ± 1.900  ns/op
DigestBenchmark.update  avgt    5  166.830 ± 4.303  ns/op

wasmJs summary:
Benchmark               Mode  Cnt    Score   Error  Units
DigestBenchmark.copy    avgt    5  107.673 ± 0.567  ns/op
DigestBenchmark.update  avgt    5   85.654 ± 0.647  ns/op
```

</details>